feature: structure for RPG systems
The initial roll comand kept all of the system-related logic in the command itself.  Slate is intended to work alternately as a Discord Bot or as an API, which means the same logic will need to be shared between bot commands and API endpoints.  Suitably, the logic for the systems needs to be shared, rather than contained in the bot command.  View-related information is still kept in the command, as the formatting of the results will be different between the command and the API.